04/02 v1 · hibernate end

+++ 1_Data_Access_2/7_Hibernate/Parte4_FuncionalidadLogin/hibernate/src/main/java/org/example/nachoHibernateConSpring/service/MedRecordService.java
@@ -2,7 +2,10 @@ package org.example.nachoHibernateConSpring.service;
 
 import org.example.nachoHibernateConSpring.dao.model.MedRecord;
 import org.example.nachoHibernateConSpring.dao.model.Medication;
+import org.example.nachoHibernateConSpring.dao.model.Session;
 import org.example.nachoHibernateConSpring.dao.repository.MedRecordDAO;
+import org.example.nachoHibernateConSpring.dao.repository.SessionDao;
+import org.example.nachoHibernateConSpring.domain.error.BadRequestException;
 import org.example.nachoHibernateConSpring.domain.model.MedRecordUI;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -17,19 +20,29 @@ public class MedRecordService {
     private final MedicationService medicationService;
     private final PatientService patientService;
     private final DoctorService doctorService;
+    private final SessionDao sessionDao;
 
-    public MedRecordService(MedRecordDAO dao, MedicationService medicationService, PatientService patientService, DoctorService doctorService) {
+    public MedRecordService(MedRecordDAO dao, MedicationService medicationService, PatientService patientService, DoctorService doctorService, SessionDao sessionDao) {
         this.dao = dao;
         this.medicationService = medicationService;
         this.patientService = patientService;
         this.doctorService = doctorService;
+        this.sessionDao = sessionDao;
     }
 
     @Transactional
     public int add(MedRecordUI medRecordUI) {
+        Session session = sessionDao.load();
+        if (session == null) throw new BadRequestException("No active session");
+
+        if (session.getUserType().equals("doctor") && medRecordUI.getIdDoctor() != session.getUserId()) {
+            throw new BadRequestException("Doctor can only add records assigned to themselves");
+        }else if (session.getUserType().equals("patient")){
+            throw new BadRequestException("Patient cannot add medical records");
+        }
+
         MedRecord medRecord = medRecordUI.toMedRecord(patientService, doctorService);
 
-        // Create and associate Medication entities
         List<Medication> medications = medRecordUI.getMedications().stream()
                 .map(medicationName -> new Medication(medicationName, medRecord))
                 .collect(Collectors.toList());
@@ -41,29 +54,55 @@ public class MedRecordService {
 
     @Transactional
     public void update(MedRecordUI medRecordUI) {
+        Session session = sessionDao.load();
+        if (session == null) throw new BadRequestException("No active session");
+
+        if (session.getUserType().equals("doctor")
+                && dao.findById(medRecordUI.getId()).get().getDoctor().getId()!=session.getUserId()){
+            throw new BadRequestException("Doctor cannot update medical records that are not assigned to them"); //this should never happen, but just in case
+        }
+
         MedRecord existingMedRecord = dao.findById(medRecordUI.getId())
                 .orElseThrow(() -> new IllegalArgumentException("MedRecord not found"));
 
         existingMedRecord.setDiagnosis(medRecordUI.getDescription());
         existingMedRecord.setDate(LocalDate.parse(medRecordUI.getDate()));
 
-        // Update Medications
         existingMedRecord.getMedications().clear();
         List<Medication> medications = medRecordUI.getMedications().stream()
                 .map(medicationName -> new Medication(medicationName, existingMedRecord))
-                .collect(Collectors.toList());
+                .toList();
         existingMedRecord.getMedications().addAll(medications);
 
         dao.save(existingMedRecord);
     }
 
     public void delete(int medRecordId) {
+        Session session = sessionDao.load();
+        if (session == null) throw new BadRequestException("No active session");
+
+        if (session.getUserType().equals("doctor")
+                && dao.findById(medRecordId).get().getDoctor().getId()!=session.getUserId()){
+            throw new BadRequestException("Doctor cannot delete medical records that are not assigned to them");//this should never happen, but just in case
+        }else if (session.getUserType().equals("patient")){
+            throw new BadRequestException("Patient cannot delete medical records");
+        }
+        
         medicationService.delete(medRecordId);
         dao.deleteById(medRecordId);
     }
 
     public List<MedRecordUI> getMedRecords(int patientId) {
-        return dao.findByPatientId(patientId).stream().map(MedRecord::toMedRecordUI).toList();
+        List<MedRecord> medRecords = dao.findByPatientId(patientId);
+        Session session = sessionDao.load();
+        if (session == null) throw new BadRequestException("No active session");
+        if (session.getUserType().equals("doctor")) {
+            medRecords = medRecords.stream()
+                    .filter(medRecord -> medRecord.getDoctor().getId() == session.getUserId())
+                    .toList();
+        }
+
+        return medRecords.stream().map(MedRecord::toMedRecordUI).toList();
     }
 
     public boolean checkPatientMedRecords(int patientId) {
